requirements.txt
pylint==2.17.5
flake8==6.1.0
isort==5.12.0
mypy==1.5.1
colorama==0.4.6
requests==2.31.0
tk==0.1.0
pyyaml==6.0.1
requests==2.31.0
